# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  clerkUserId: string;
  displayName: string;
  jobRole: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  preferredDays?: string[];
  preferredTimes?: string[];
  /** @nullable */
  minHourlyRate?: number | null;
  /** @nullable */
  maxHourlyRate?: number | null;
  isPro?: boolean;
  createdAt: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  displayName: string;
  /** @minLength 1 */
  jobRole: string;
  location?: string;
  bio?: string;
  avatarUrl?: string;
  preferredDays?: string[];
  preferredTimes?: string[];
  minHourlyRate?: number;
  maxHourlyRate?: number;
}

export type ShiftShiftType = typeof ShiftShiftType[keyof typeof ShiftShiftType];


export const ShiftShiftType = {
  swap: 'swap',
  cover: 'cover',
} as const;

export type ShiftStatus = typeof ShiftStatus[keyof typeof ShiftStatus];


export const ShiftStatus = {
  open: 'open',
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
} as const;

export interface Shift {
  id: number;
  clerkUserId: string;
  title: string;
  role: string;
  location: string;
  shiftDate: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  hourlyRate?: number | null;
  shiftType: ShiftShiftType;
  /** @nullable */
  notes?: string | null;
  status: ShiftStatus;
  createdAt: string;
}

export type ShiftWithProfileShiftType = typeof ShiftWithProfileShiftType[keyof typeof ShiftWithProfileShiftType];


export const ShiftWithProfileShiftType = {
  swap: 'swap',
  cover: 'cover',
} as const;

export type ShiftWithProfileStatus = typeof ShiftWithProfileStatus[keyof typeof ShiftWithProfileStatus];


export const ShiftWithProfileStatus = {
  open: 'open',
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
} as const;

export interface ShiftWithProfile {
  id: number;
  clerkUserId: string;
  title: string;
  role: string;
  location: string;
  shiftDate: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  hourlyRate?: number | null;
  shiftType: ShiftWithProfileShiftType;
  /** @nullable */
  notes?: string | null;
  status: ShiftWithProfileStatus;
  createdAt: string;
  poster?: Profile;
  requestCount?: number;
}

export type ShiftInputShiftType = typeof ShiftInputShiftType[keyof typeof ShiftInputShiftType];


export const ShiftInputShiftType = {
  swap: 'swap',
  cover: 'cover',
} as const;

export interface ShiftInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  location: string;
  shiftDate: string;
  startTime: string;
  endTime: string;
  hourlyRate?: number;
  shiftType: ShiftInputShiftType;
  notes?: string;
}

export type ShiftUpdateShiftType = typeof ShiftUpdateShiftType[keyof typeof ShiftUpdateShiftType];


export const ShiftUpdateShiftType = {
  swap: 'swap',
  cover: 'cover',
} as const;

export type ShiftUpdateStatus = typeof ShiftUpdateStatus[keyof typeof ShiftUpdateStatus];


export const ShiftUpdateStatus = {
  open: 'open',
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
} as const;

export interface ShiftUpdate {
  title?: string;
  role?: string;
  location?: string;
  shiftDate?: string;
  startTime?: string;
  endTime?: string;
  hourlyRate?: number;
  shiftType?: ShiftUpdateShiftType;
  notes?: string;
  status?: ShiftUpdateStatus;
}

export type ShiftRequestStatus = typeof ShiftRequestStatus[keyof typeof ShiftRequestStatus];


export const ShiftRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ShiftRequest {
  id: number;
  shiftId: number;
  clerkUserId: string;
  /** @nullable */
  message?: string | null;
  status: ShiftRequestStatus;
  createdAt: string;
}

export interface ShiftRequestInput {
  message?: string;
}

export type ShiftRequestWithProfileStatus = typeof ShiftRequestWithProfileStatus[keyof typeof ShiftRequestWithProfileStatus];


export const ShiftRequestWithProfileStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ShiftRequestWithProfile {
  id: number;
  shiftId: number;
  clerkUserId: string;
  /** @nullable */
  message?: string | null;
  status: ShiftRequestWithProfileStatus;
  createdAt: string;
  requester?: Profile;
}

export type ShiftRequestWithShiftStatus = typeof ShiftRequestWithShiftStatus[keyof typeof ShiftRequestWithShiftStatus];


export const ShiftRequestWithShiftStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ShiftRequestWithShift {
  id: number;
  shiftId: number;
  clerkUserId: string;
  /** @nullable */
  message?: string | null;
  status: ShiftRequestWithShiftStatus;
  createdAt: string;
  shift?: Shift;
}

export interface Message {
  id: number;
  conversationId: string;
  senderClerkUserId: string;
  content: string;
  createdAt: string;
  senderProfile?: Profile;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ConversationSummary {
  conversationId: string;
  otherUser: Profile;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface DashboardStats {
  myActiveShifts: number;
  incomingRequests: number;
  outgoingRequests: number;
  totalSwapsCompleted: number;
  isPro: boolean;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  shift_posted: 'shift_posted',
  request_received: 'request_received',
  request_approved: 'request_approved',
  request_rejected: 'request_rejected',
  swap_completed: 'swap_completed',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  createdAt: string;
  /** @nullable */
  shiftId?: number | null;
}

export interface CheckoutSessionInput {
  successUrl: string;
  cancelUrl: string;
}

export interface CheckoutSession {
  url: string;
}

export interface CheckoutVerification {
  success: boolean;
  isPro: boolean;
}

export type ListShiftsParams = {
/**
 * @nullable
 */
role?: string | null;
/**
 * @nullable
 */
location?: string | null;
/**
 * @nullable
 */
shiftType?: string | null;
/**
 * @nullable
 */
dateFrom?: string | null;
/**
 * @nullable
 */
dateTo?: string | null;
/**
 * @nullable
 */
minPay?: number | null;
/**
 * @nullable
 */
maxPay?: number | null;
/**
 * @nullable
 */
mine?: boolean | null;
};

export type VerifyCheckoutParams = {
session_id: string;
};