# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShiftSwap API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    put:
      operationId: upsertProfile
      tags: [profiles]
      summary: Create or update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /requests/{id}/approve:
    patch:
      operationId: approveRequest
      tags: [requests]
      summary: Approve a shift request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftRequest"
  /requests/{id}/reject:
    patch:
      operationId: rejectRequest
      tags: [requests]
      summary: Reject a shift request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftRequest"
...
  /messages/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /checkout/create-session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create Stripe checkout session for Pro Lifetime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /checkout/verify:
    get:
      operationId: verifyCheckout
      tags: [checkout]
      summary: Verify checkout session and upgrade user to Pro
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutVerification"
components:
  schemas:
...
    Profile:
      type: object
      required: [id, clerkUserId, displayName, jobRole, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        displayName:
          type: string
        jobRole:
          type: string
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        preferredDays:
          type: array
          items:
            type: string
        preferredTimes:
          type: array
          items:
            type: string
        minHourlyRate:
          type: ["number", "null"]
        maxHourlyRate:
          type: ["number", "null"]
        isPro:
          type: boolean
        createdAt:
          type: string
    ProfileInput:
      type: object
      required: [displayName, jobRole]
      properties:
        displayName:
          type: string
          minLength: 1
        jobRole:
          type: string
          minLength: 1
        location:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        preferredDays:
          type: array
          items:
            type: string
        preferredTimes:
          type: array
          items:
            type: string
        minHourlyRate:
          type: number
        maxHourlyRate:
          type: number
...
    ShiftRequest:
      type: object
      required: [id, shiftId, clerkUserId, status, createdAt]
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        clerkUserId:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
...
    Message:
      type: object
      required: [id, conversationId, senderClerkUserId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: string
        senderClerkUserId:
          type: string
        content:
          type: string
        createdAt:
          type: string
        senderProfile:
          $ref: "#/components/schemas/Profile"
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...
    CheckoutSessionInput:
      type: object
      required: [successUrl, cancelUrl]
      properties:
        successUrl:
          type: string
        cancelUrl:
          type: string
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
    CheckoutVerification:
      type: object
      required: [success, isPro]
      properties:
        success:
          type: boolean
        isPro:
          type: boolean