# lib/api-zod/src/generated/types/checkoutVerification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutVerification {
  success: boolean;
  isPro: boolean;
}