# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}