# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  clerkUserId: string;
  displayName: string;
  jobRole: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  preferredDays?: string[];
  preferredTimes?: string[];
  /** @nullable */
  minHourlyRate?: number | null;
  /** @nullable */
  maxHourlyRate?: number | null;
  isPro?: boolean;
  createdAt: string;
}